requirements.txt
python-telegram-bot==13.7
PySocks==1.7.1
redis==3.5.3